Return warnings from `validate`
